[refactor] reorganising parser packages
