+++ lib/data/populartimes/payloads/dayResults.dart
